requirements.txt
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cycler==0.11.0
Flask==2.2.2
fonttools==4.38.0
idna==3.4
importlib-metadata==5.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jsonify==0.5
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.3
numpy==1.21.6
packaging==22.0
pandas==1.3.5
Pillow==9.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.1
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
threadpoolctl==3.1.0
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2
wincertstore==0.2
zipp==3.11.0
